# docs/swagger.yaml
basePath: /
definitions:
  dto.CompanyRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.DepartmentRequest:
    properties:
      company_id:
        type: string
      department_name:
        items:
          type: string
        type: array
    type: object
  dto.FormRequest:
    properties:
      description:
        type: string
      link_id:
        type: string
      title:
        type: string
    type: object
  dto.LinkRequest:
    properties:
      company_id:
        type: string
    type: object
  dto.TicketRequest:
    properties:
      department_id:
        type: string
      description:
        type: string
      form_id:
        type: string
      message:
        type: string
      priority:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8888
info:
  contact: {}
  description: CTD Backend API
  title: Predict Ticket Department API
  version: "1.0"
paths:
  /api/v1/company_form/{company_id}:
    get:
      consumes:
      - application/json
      description: Get Company Form By company ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Company Form By company ID
      tags:
      - company_form
  /api/v1/company_form/create:
    post:
      consumes:
      - application/json
      description: Create Company Form
      parameters:
      - description: Link credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Company Form
      tags:
      - company_form
  /api/v1/departments/add:
    post:
      consumes:
      - application/json
      description: Create department in company
      parameters:
      - description: Department credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add Departments
      tags:
      - department
  /api/v1/departments/company/{company_id}:
    get:
      consumes:
      - application/json
      description: Get Departments By company ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Departments By company ID
      tags:
      - department
  /api/v1/forms/{company_id}:
    get:
      consumes:
      - application/json
      description: Get all forms By company ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Forms By company ID
      tags:
      - form
  /api/v1/forms/{company_id}/per-day:
    get:
      consumes:
      - application/json
      description: Get all forms per day By company ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      - description: Date in YYYY-MM-DD format; defaults to yesterday
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Forms Per Day By company ID
      tags:
      - form
  /api/v1/forms/submit:
    post:
      consumes:
      - application/json
      description: Submit Form to Company
      parameters:
      - description: Form credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit Form
      tags:
      - form
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate and get JWT token
      parameters:
      - description: Company credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login company
      tags:
      - Company
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create new company accout
      parameters:
      - description: Company credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register company
      tags:
      - Company
  /api/v1/tickets/{company_id}:
    get:
      consumes:
      - application/json
      description: Get all tickets by company ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Tickets By company ID
      tags:
      - ticket
  /api/v1/tickets/create:
    post:
      consumes:
      - application/json
      description: Create a single ticket
      parameters:
      - description: Ticket credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Ticket
      tags:
      - ticket
  /api/v1/tickets/create-bulk:
    post:
      consumes:
      - application/json
      description: Create multiple tickets in one request
      parameters:
      - description: Ticket credentials
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.TicketRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Tickets
      tags:
      - ticket
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"